# mint-ensemble-manager/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mint Ensemble Manager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://portals.tapis.io/v3/oauth2/authorize",
            "scopes": {}
          }
        }
      }
    },
    "schemas": {
      "TimePeriod": {
        "type": "object",
        "description": "A time period",
        "properties": {
          "start_date": {
            "description": "From date YYYY-MM-DD format",
            "type": "string",
            "format": "date-time"
          },
          "end_date": {
            "description": "to date YYYY-MM-DD format",
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MintPermission": {
        "type": "object",
        "properties": {
          "userid": {
            "type": "string",
            "example": "*"
          },
          "read": {
            "type": "boolean"
          },
          "write": {
            "type": "boolean"
          },
          "execute": {
            "type": "boolean"
          },
          "owner": {
            "type": "boolean"
          }
        }
      },
      "MintEvent": {
        "type": "object",
        "properties": {
          "event": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "ADD_TASK",
              "DELETE_TASK"
            ]
          },
          "userid": {
            "type": "string",
            "example": "*"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
      "Task": {
        "type": "object",
        "description": "A Task definition",
        "properties": {
          "id": {
            "description": "The id of the task (leave empty if creating a new task)",
            "type": "string"
          },
          "name": {
            "description": "The name of the task",
            "type": "string"
          },
          "indicatorid": {
            "description": "The SVO label for the indicator/response (link to list upcoming). Example: grain~dry__mass-per-area_yield",
            "type": "string",
            "example": "crop__potential_transpiration_volume_flux"
          },
          "interventionid": {
            "description": "The SVO label for the intervention (if any). Example: crop__planting_start_time",
            "type": "string",
            "example": "nitrogen__average_of_net_mass_mineralization_rate"
          },
          "regionid": {
            "description": "The specific region id for the task (browse https://dev.mint.isi.edu/ethiopia/regions to get the ids)",
            "type": "string"
          },
          "dates": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "events": {
            "type": "array",
            "description": "List of events associated with the task",
            "items": {
              "$ref": "#/components/schemas/MintEvent"
            }
          },
          "permissions": {
            "type": "array",
            "description": "List of permissions for the task",
            "items": {
              "$ref": "#/components/schemas/MintPermission"
            }
          }
        }
      },
      "CreateTaskRequest": {
        "type": "object",
        "description": "A Task creation request definition",
        "required": [
          "name",
          "dates"
        ],
        "properties": {
          "name": {
            "description": "The name of the task",
            "type": "string",
            "example": "Texas Water Management Analysis 2024"
          },
          "dates": {
            "type": "object",
            "description": "The date range for the task",
            "properties": {
              "start_date": {
                "description": "Start date in ISO format",
                "type": "string",
                "format": "date-time",
                "example": "2024-01-01T00:00:00Z"
              },
              "end_date": {
                "description": "End date in ISO format",
                "type": "string",
                "format": "date-time",
                "example": "2024-12-31T23:59:59Z"
              }
            }
          },
          "response_variables": {
            "example": [
              "crop__potential_transpiration_volume_flux"
            ],
            "type": "array",
            "description": "List of response variables",
            "items": {
              "type": "string"
            }
          },
          "driving_variables": {
            "example": [
              "nitrogen__average_of_net_mass_mineralization_rate"
            ],
            "type": "array",
            "description": "List of driving variables",
            "items": {
              "type": "string"
            }
          },
          "regionid": {
            "description": "The specific region id for the task",
            "type": "string",
            "example": "texas"
          },
          "events": {
            "type": "array",
            "description": "List of events associated with the task",
            "items": {
              "$ref": "#/components/schemas/MintEvent"
            }
          },
          "permissions": {
            "type": "array",
            "description": "List of permissions for the task",
            "items": {
              "$ref": "#/components/schemas/MintPermission"
            }
          }
        }
      },
...
      "CreateSubtaskRequest": {
        "type": "object",
        "description": "A Subtask creation request definition",
        "required": [
          "name",
          "dates",
          "task_id",
          "driving_variables",
          "response_variables"
        ],
        "properties": {
          "name": {
            "description": "The name of the subtask",
            "type": "string"
          },
          "dates": {
            "type": "object",
            "description": "The date range for the subtask",
            "properties": {
              "start_date": {
                "description": "Start date in ISO format",
                "type": "string",
                "format": "date-time"
              },
              "end_date": {
                "description": "End date in ISO format",
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "task_id": {
            "description": "The ID of the parent task",
            "type": "string"
          },
          "driving_variables": {
            "type": "array",
            "description": "List of driving variables",
            "items": {
              "type": "string"
            }
          },
          "response_variables": {
            "type": "array",
            "description": "List of response variables",
            "items": {
              "type": "string"
            }
          },
          "regionid": {
            "description": "The specific region id for the subtask",
            "type": "string"
          },
          "events": {
            "type": "array",
            "description": "List of events associated with the subtask",
            "items": {
              "$ref": "#/components/schemas/MintEvent"
            }
          },
          "permissions": {
            "type": "array",
            "description": "List of permissions for the subtask",
            "items": {
              "$ref": "#/components/schemas/MintPermission"
            }
          }
        }
      },
      "Subtask": {
        "type": "object",
        "description": "A Subtask definition",
        "properties": {
          "name": {
            "description": "The name of the thread (optional)",
            "type": "string"
          },
          "modelid": {
            "description": "The model id to use (browse here: https://dev.mint.isi.edu/ethiopia/models/explore. Example: cycles-0.10.2-alpha-collection-oromia)",
            "type": "string"
          },
          "datasets": {
            "description": "A map of input name and dataset ids. Example { cycles_weather_soil: ac34f01b-1484-4403-98ea-3a380838cab1 }. To browse the datasets, go here: http://localhost:8080/ethiopia/datasets/browse",
            "type": "object"
          },
          "parameters": {
            "description": "A map of parameter name and parameter values. Example: { start_planting_day: [100, 107, 114] }",
            "type": "object"
          }
        }
      },
...
      "ModelThread": {
        "description": "",
        "required": [
          "thread_id"
        ],
        "type": "object",
        "properties": {
          "thread_id": {
            "description": "The Modeling thread id",
            "type": "string"
          },
          "model_id": {
            "description": "The Model id (optional. Set to null to run all)",
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/registration": {
      "post": {
        "summary": "Register outputs of modeling thread in data catalog",
        "operationId": "registerExecutionOutputs",
        "tags": [
          "Registration"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "oauth2": []
          }
        ],
        "requestBody": {
          "description": "Modeling thread scenario/subgoal/id",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelThread"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Execution outputs registered successfully"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/problemStatements/{problemStatementId}/tasks": {
      "get": {
        "summary": "Get all tasks for a problem statement",
        "description": "Returns a list of all tasks associated with a specific problem statement",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "problemStatementId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The problem statement ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "description": "Creates a new task for a specific problem statement",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "problemStatementId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The problem statement ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/problemStatements/{problemStatementId}/tasks/{taskId}/subtasks": {
      "get": {
        "summary": "Get all subtasks for a task",
        "description": "Returns a list of all subtasks associated with a specific task",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Subtasks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "problemStatementId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The problem statement ID"
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The task ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of subtasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subtasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Subtask"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new subtask",
        "description": "Creates a new subtask for a specific task",
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "tags": [
          "Subtasks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "problemStatementId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The problem statement ID"
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The task ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubtaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subtask created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subtask"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/tapis/executions/{executionId}/outputs": {
      "post": {
        "summary": "Register Tapis Execution Outputs",
        "description": "Register the outputs of a successful Tapis execution in the data catalog",
        "operationId": "registerTapisExecutionOutputs",
        "tags": [
          "Tapis"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "name": "executionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Outputs registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or registration failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/tapis/modelEnsembles/{id}/outputs": {
      "post": {
        "summary": "Register Model Ensemble Execution Outputs",
        "description": "Register all execution outputs for a model ensemble in the data catalog asynchronously",
        "operationId": "registerModelEnsembleExecutionOutputs",
        "tags": [
          "Tapis"
        ],
        "security": [
          {
            "BearerAuth": [],
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Registration process started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}